Use WCIF model class consequently
