+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/wcif/model/Venue.kt
@@ -1,7 +1,5 @@
 package org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model
 
-import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.Room
-import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.SafeNamed
 import java.time.ZoneId
 
 data class Venue(val name: String, val rooms: List<Room>, val timezone: String) : SafeNamed(name) {
