+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/wcif/model/WCIF.kt
@@ -1,0 +1,3 @@
+package org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model
+
+data class WCIF(val schedule: Schedule)
